Add support for multiple operating systems
Using a new input field in the json credentials file, the user can now specify the operating system of their server. Previously it was assumed that the server was always linux/unix.

The application uses this input to differentiate what commands it should execute over SSH to extract that server's performance data.

Supported options are Linux, Windows, and MacOS, if nothing is specified the application assumes the server is linux/unix